Refactor row/col specific code into general methods
Many row- and column-specific methods are basically copies of
each other, with "row" and "column" switched.  Instead of copying
the code and switching "row" and "column", private methods were
created that abstract away the "row" and "column" specifics, so
the row- and column-specific methods can call these private
methods, passing in whether it's for rows or columns.  This
reduces code duplication.